test(cgroups): add code coverage for job annotation transform plugin
