+++ plugins/cgroups/util-cgroups-plugins/src/lib.rs
@@ -10,4 +10,3 @@ pub mod job_annotation_transform;
 pub mod metrics;
 pub mod regex;
 mod self_stop;
-mod tests;
